Create a Profiles/ subfolder inside the components/ folder Move AdminProfile into the Profiles/ subfolder Move UserProfile into the Profiles/ subfolder Modify paths in affected components that import these two components Modify paths in these two components that import other components
