+++ client/components/Profiles/UserProfile.jsx
@@ -3,10 +3,10 @@ import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import { bindActionCreators } from 'redux';
-import Search from './common/Search';
-import EventsListWithImage from './events/presentational/EventsListWithImage';
-import Preloader from './common/Preloader';
-import * as eventActions from '../actions/eventActions';
+import Search from '../common/Search';
+import EventsListWithImage from '../events/presentational/EventsListWithImage';
+import Preloader from '../common/Preloader';
+import * as eventActions from '../../actions/eventActions';
 
 class UserProfile extends Component {
   constructor(props) {
